Changes related for: - Some methods was working with integer for number of elements in a TPZVec - Some methods had arguments no ready to STATE, this changes avoid many warnings - Clone classes was modified but they need validation yet.

+++ External/pzmetis.h
@@ -20,7 +20,7 @@ public:
 	 * @brief Perform the renumbering of elements. The aim of this operation is to minimize the
 	 * band of the resulting stiffeness matrix.
 	 */
-	virtual void Resequence(TPZVec<int> &perm, TPZVec<int> &inverseperm);
+	virtual void Resequence(TPZVec<long> &perm, TPZVec<long> &inverseperm);
 	
     /** @brief Constructor. */
 	/**
